I kind of forget all the stuff (sorry)
I added a game manager that keeps track of what level you're on.

I set up the next level button to load the next level (it only works on
the first level thought -- doesn't show up on the second).

Added developer shortcuts so you can hit 1, 2 or 3 to go to the
different levels.

Maybe a smattering of other stuff...Oh, and commenting the main scripts
(algaeLifespan and gameManager)